Added Pinning Collections and Simple Build Sidebar
This commit adds the ability to pin collections, allow a user to save public collections to a list for easy access.
It also moves the build viewing feature from a seperate page to a sidebar that pops out when clicked.
This allows easier access to viewing ongoing builds while browsing the website.
This feature is still incomplete and buggy, but it should work as well as the old build page for all practical functions.
(If you need to access the old page, you can still do it at #/build/<buildId>)

Please take note that the database was modified in this update, and you must update any local database to run this.
For instructions on how to do this, please see the docs in the master branch.