Improved build tracker and add basic Public Collection search.
The build tracker should now automatically display running builds when loaded, and builds can be manually shown or hidden.
Its tabs now also serve as progress bars, to allow easy tracking of progress when they are closed.
The search function on the Public Collection panel is currently very simple, and only checks if the search term is a substring in the collection title or description.

Please note that the old build viewer is now no longer implemented.