Broken Hermes demo
Made many major changes, but those are going to be described in a commit that works and is finished.
This commit is to test if other people are having the same problem I'm having, which is that the .impl package in the Hermes Filter project isn't being linked for no apparent reason.